move sending channel message handling to channel store.
